requirements.txt
torch>=1.10.0
transformers>=4.16.0
pdf2image>=1.16.0
pytesseract>=0.3.8
opencv-python>=4.5.5
pandas>=1.3.5
openpyxl>=3.0.9
Pillow>=9.0.0
reportlab>=3.6.0
img2table>=1.0.0
streamlit>=1.22.0
langgraph>=0.0.19
openai>=1.3.0
typing-extensions>=4.11.0
python-dotenv>=1.0.0
